+++ src/configuration/from_file.ts
@@ -1,24 +1,31 @@
 import stringArgv from 'string-argv'
+import fs from 'fs'
 import path from 'path'
+import YAML from 'yaml'
+import { promisify } from 'util'
 import { pathToFileURL } from 'url'
 import { IConfiguration } from './types'
 import { mergeConfigurations } from './merge_configurations'
 import ArgvParser from './argv_parser'
 import { checkSchema } from './check_schema'
+import { ILogger } from '../logger'
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
 const { importer } = require('../importer')
 
 export async function fromFile(
+  logger: ILogger,
   cwd: string,
   file: string,
   profiles: string[] = []
 ): Promise<Partial<IConfiguration>> {
   const definitions = await loadFile(cwd, file)
   if (!definitions.default) {
+    logger.debug('No default profile defined in configuration file')
     definitions.default = {}
   }
   if (profiles.length < 1) {
+    logger.debug('No profiles specified; using default profile')
     profiles = ['default']
   }
   const definedKeys = Object.keys(definitions)
@@ -30,7 +37,7 @@ export async function fromFile(
   return mergeConfigurations(
     {},
     ...profiles.map((profileKey) =>
-      extractConfiguration(profileKey, definitions[profileKey])
+      extractConfiguration(logger, profileKey, definitions[profileKey])
     )
   )
 }
@@ -40,16 +47,31 @@ async function loadFile(
   file: string
 ): Promise<Record<string, any>> {
   const filePath: string = path.join(cwd, file)
+  const extension = path.extname(filePath)
   let definitions
-  try {
-    // eslint-disable-next-line @typescript-eslint/no-var-requires
-    definitions = require(filePath)
-  } catch (error) {
-    if (error.code === 'ERR_REQUIRE_ESM') {
-      definitions = await importer(pathToFileURL(filePath))
-    } else {
-      throw error
-    }
+  switch (extension) {
+    case '.json':
+      definitions = JSON.parse(
+        await promisify(fs.readFile)(filePath, { encoding: 'utf-8' })
+      )
+      break
+    case '.yaml':
+    case '.yml':
+      definitions = YAML.parse(
+        await promisify(fs.readFile)(filePath, { encoding: 'utf-8' })
+      )
+      break
+    default:
+      try {
+        // eslint-disable-next-line @typescript-eslint/no-var-requires
+        definitions = require(filePath)
+      } catch (error) {
+        if (error.code === 'ERR_REQUIRE_ESM') {
+          definitions = await importer(pathToFileURL(filePath))
+        } else {
+          throw error
+        }
+      }
   }
   if (typeof definitions !== 'object') {
     throw new Error(`Configuration file ${filePath} does not export an object`)
@@ -58,10 +80,12 @@ async function loadFile(
 }
 
 function extractConfiguration(
+  logger: ILogger,
   name: string,
   definition: any
 ): Partial<IConfiguration> {
   if (typeof definition === 'string') {
+    logger.debug(`Profile "${name}" value is a string; parsing as argv`)
     const { configuration } = ArgvParser.parse([
       'node',
       'cucumber-js',