+++ src/runtime/test_case_runner_spec.ts
@@ -12,7 +12,7 @@ import timeMethods from '../time'
 import { getBaseSupportCodeLibrary } from '../../test/fixtures/steps'
 import { ISupportCodeLibrary } from '../support_code_library_builder/types'
 import { valueOrDefault } from '../value_checker'
-import { PredictableTestRunStopwatch } from './stopwatch'
+import { create } from './stopwatch'
 import { assembleTestCases } from './assemble_test_cases'
 import IEnvelope = messages.Envelope
 
@@ -48,12 +48,13 @@ async function testRunner(
   eventBroadcaster.on('envelope', (e) => envelopes.push(e))
   const runner = new TestCaseRunner({
     eventBroadcaster,
-    stopwatch: new PredictableTestRunStopwatch(),
+    stopwatch: create(),
     gherkinDocument: options.gherkinDocument,
     newId,
     pickle: options.pickle,
     testCase,
     retries: valueOrDefault(options.retries, 0),
+    filterStackTraces: false,
     skip: valueOrDefault(options.skip, false),
     supportCodeLibrary: options.supportCodeLibrary,
     worldParameters: {},
@@ -110,7 +111,7 @@ describe('TestCaseRunner', () => {
         })
 
         // Assert
-        const expectedtEnvelopes: messages.Envelope[] = [
+        const expectedEnvelopes: messages.Envelope[] = [
           {
             testCaseStarted: {
               attempt: 0,
@@ -123,7 +124,7 @@ describe('TestCaseRunner', () => {
             testStepStarted: {
               testCaseStartedId: '2',
               testStepId: '1',
-              timestamp: predictableTimestamp(1),
+              timestamp: predictableTimestamp(0),
             },
           },
           {
@@ -131,18 +132,18 @@ describe('TestCaseRunner', () => {
               testCaseStartedId: '2',
               testStepResult: passedTestResult,
               testStepId: '1',
-              timestamp: predictableTimestamp(2),
+              timestamp: predictableTimestamp(1),
             },
           },
           {
             testCaseFinished: {
               testCaseStartedId: '2',
-              timestamp: predictableTimestamp(3),
+              timestamp: predictableTimestamp(1),
               willBeRetried: false,
             },
           },
         ]
-        expect(envelopes).to.eql(expectedtEnvelopes)
+        expect(envelopes).to.eql(expectedEnvelopes)
         expect(result).to.eql(messages.TestStepResultStatus.PASSED)
       })
     })
@@ -262,6 +263,7 @@ describe('TestCaseRunner', () => {
         const supportCodeLibrary = buildSupportCodeLibrary(({ Given }) => {
           let willPass = false
           Given('a step', function () {
+            clock.tick(1)
             if (willPass) {
               return
             }
@@ -299,25 +301,25 @@ describe('TestCaseRunner', () => {
             testStepStarted: {
               testCaseStartedId: '2',
               testStepId: '1',
-              timestamp: predictableTimestamp(1),
+              timestamp: predictableTimestamp(0),
             },
           },
           {
             testStepFinished: {
               testCaseStartedId: '2',
               testStepResult: {
-                duration: messages.TimeConversion.millisecondsToDuration(0),
+                duration: messages.TimeConversion.millisecondsToDuration(1),
                 message: 'error',
                 status: messages.TestStepResultStatus.FAILED,
               },
               testStepId: '1',
-              timestamp: predictableTimestamp(2),
+              timestamp: predictableTimestamp(1),
             },
           },
           {
             testCaseFinished: {
               testCaseStartedId: '2',
-              timestamp: predictableTimestamp(3),
+              timestamp: predictableTimestamp(1),
               willBeRetried: true,
             },
           },
@@ -326,32 +328,32 @@ describe('TestCaseRunner', () => {
               attempt: 1,
               id: '3',
               testCaseId: '0',
-              timestamp: predictableTimestamp(4),
+              timestamp: predictableTimestamp(1),
             },
           },
           {
             testStepStarted: {
               testCaseStartedId: '3',
               testStepId: '1',
-              timestamp: predictableTimestamp(5),
+              timestamp: predictableTimestamp(1),
             },
           },
           {
             testStepFinished: {
               testCaseStartedId: '3',
               testStepResult: {
-                duration: messages.TimeConversion.millisecondsToDuration(0),
+                duration: messages.TimeConversion.millisecondsToDuration(1),
                 message: undefined,
                 status: messages.TestStepResultStatus.PASSED,
               },
               testStepId: '1',
-              timestamp: predictableTimestamp(6),
+              timestamp: predictableTimestamp(2),
             },
           },
           {
             testCaseFinished: {
               testCaseStartedId: '3',
-              timestamp: predictableTimestamp(7),
+              timestamp: predictableTimestamp(2),
               willBeRetried: false,
             },
           },