+++ src/time.ts
@@ -1,25 +1,21 @@
 import { performance } from 'perf_hooks'
 import * as messages from '@cucumber/messages'
-import { FakeClock, GlobalTimers, TimerId } from '@sinonjs/fake-timers'
 
-let previousTimestamp: number
-
-interface IMethods extends GlobalTimers<TimerId> {
-  beginTiming: () => void
-  endTiming: () => number
-  performance: FakeClock<TimerId>['performance']
+interface ProtectedTimingBuiltins {
+  clearImmediate: typeof clearImmediate
+  clearInterval: typeof clearInterval
+  clearTimeout: typeof clearTimeout
+  Date: typeof Date
+  setImmediate: typeof setImmediate
+  setInterval: typeof setInterval
+  setTimeout: typeof setTimeout
+  performance: typeof performance
 }
 
-const methods: Partial<IMethods> = {
-  beginTiming() {
-    previousTimestamp = getTimestamp()
-  },
+const methods: Partial<ProtectedTimingBuiltins> = {
   clearInterval: clearInterval.bind(global),
   clearTimeout: clearTimeout.bind(global),
   Date,
-  endTiming() {
-    return getTimestamp() - previousTimestamp
-  },
   setInterval: setInterval.bind(global),
   setTimeout: setTimeout.bind(global),
   performance,
@@ -28,10 +24,6 @@ const methods: Partial<IMethods> = {
 if (typeof setImmediate !== 'undefined') {
   methods.setImmediate = setImmediate.bind(global)
   methods.clearImmediate = clearImmediate.bind(global)
-}
-
-function getTimestamp(): number {
-  return methods.performance.now()
 }
 
 export function durationBetweenTimestamps(
@@ -51,7 +43,7 @@ export async function wrapPromiseWithTimeout<T>(
   timeoutInMilliseconds: number,
   timeoutMessage: string = ''
 ): Promise<T> {
-  let timeoutId: TimerId
+  let timeoutId: ReturnType<typeof setTimeout>
   if (timeoutMessage === '') {
     timeoutMessage = `Action did not complete within ${timeoutInMilliseconds} milliseconds`
   }