+++ src/wrapper.mjs
@@ -1,9 +1,5 @@
 import cucumber from './index.js'
 
-export const Cli = cucumber.Cli
-export const parseGherkinMessageStream = cucumber.parseGherkinMessageStream
-export const PickleFilter = cucumber.PickleFilter
-export const Runtime = cucumber.Runtime
 export const supportCodeLibraryBuilder = cucumber.supportCodeLibraryBuilder
 export const Status = cucumber.Status
 export const DataTable = cucumber.DataTable
@@ -28,10 +24,6 @@ export const Before = cucumber.Before
 export const BeforeAll = cucumber.BeforeAll
 export const BeforeStep = cucumber.BeforeStep
 export const defineParameterType = cucumber.defineParameterType
-/**
- * @deprecated use `Given`, `When` or `Then` instead; see <https://github.com/cucumber/cucumber-js/issues/2043>
- */
-export const defineStep = cucumber.defineStep
 export const Given = cucumber.Given
 export const setDefaultTimeout = cucumber.setDefaultTimeout
 export const setDefinitionFunctionWrapper = cucumber.setDefinitionFunctionWrapper
@@ -43,3 +35,9 @@ export const parallelCanAssignHelpers = cucumber.parallelCanAssignHelpers
 
 export const wrapPromiseWithTimeout = cucumber.wrapPromiseWithTimeout
 
+// Deprecated
+export const Cli = cucumber.Cli
+export const defineStep = cucumber.defineStep
+export const parseGherkinMessageStream = cucumber.parseGherkinMessageStream
+export const PickleFilter = cucumber.PickleFilter
+export const Runtime = cucumber.Runtime