+++ CHANGELOG.md
@@ -9,6 +9,37 @@ Please see [CONTRIBUTING.md](https://github.com/cucumber/cucumber/blob/master/CO
 
 ## [Unreleased]
 
+## [8.10.0] - 2022-12-27
+### Added
+- Add support for YAML as a configuration file format ([#2199](https://github.com/cucumber/cucumber-js/pull/2199))
+
+### Changed
+- Replace `duration` with `luxon` for formatting durations ([#2204](https://github.com/cucumber/cucumber-js/pull/2204))
+
+## [8.9.1] - 2022-12-16
+### Fixed
+- Include original coordinates in `loadSupport` result ([#2197](https://github.com/cucumber/cucumber-js/pull/2197))
+
+## [8.9.0] - 2022-11-24
+### Added
+- Add new JUnit formatter (see [documentation](./docs/formatters.md#junit)) ([#2121](https://github.com/cucumber/cucumber-js/pull/2121))
+
+## [8.8.0] - 2022-11-14
+### Changed
+- Add `workerId` property to `testCaseStarted` message ([#2085](https://github.com/cucumber/cucumber-js/pull/2085))
+- Handle stack traces without V8-specific modification ([#2119](https://github.com/cucumber/cucumber-js/pull/2119))
+
+## [8.7.0] - 2022-10-17
+### Deprecated
+- `Cli`, `PickleFilter` and `Runtime` deprecated in favour of new API functions (see [documentation](./docs/deprecations.md)) ([#2136](https://github.com/cucumber/cucumber-js/pull/2136))
+
+## [8.6.0] - 2022-09-20
+### Added
+- Debug logging capability to help diagnose configuration issues (see [documentation](./docs/debugging.md)) ([#2120](https://github.com/cucumber/cucumber-js/pull/2120))
+
+### Fixed
+- Rework time interfaces to avoid using sinon types ([#2142](https://github.com/cucumber/cucumber-js/pull/2142))
+
 ## [8.5.3] - 2022-09-10
 ### Fixed
 - Default `stderr` if not provided to `Cli` constructor ([#2138](https://github.com/cucumber/cucumber-js/pull/2138))
@@ -1428,7 +1459,13 @@ this.Given(), this.When(), this.Then() and this.defineStep() ([#2](https://githu
 
 ## 0.0.1
 
-[Unreleased]: https://github.com/cucumber/cucumber-js/compare/v8.5.3...HEAD
+[Unreleased]: https://github.com/cucumber/cucumber-js/compare/v8.10.0...HEAD
+[8.10.0]: https://github.com/cucumber/cucumber-js/compare/v8.9.1...v8.10.0
+[8.9.1]: https://github.com/cucumber/cucumber-js/compare/v8.9.0...v8.9.1
+[8.9.0]: https://github.com/cucumber/cucumber-js/compare/v8.8.0...v8.9.0
+[8.8.0]: https://github.com/cucumber/cucumber-js/compare/v8.7.0...v8.8.0
+[8.7.0]: https://github.com/cucumber/cucumber-js/compare/v8.6.0...v8.7.0
+[8.6.0]: https://github.com/cucumber/cucumber-js/compare/v8.5.3...v8.6.0
 [8.5.3]: https://github.com/cucumber/cucumber-js/compare/v8.5.2...v8.5.3
 [8.5.2]: https://github.com/cucumber/cucumber-js/compare/v8.5.1...v8.5.2
 [8.5.1]: https://github.com/cucumber/cucumber-js/compare/v8.5.0...v8.5.1